chore: move components to prividium folder
