+++ packages/app/src/components/ConnectMetamaskButton.vue
@@ -34,7 +34,7 @@ import { computed, ref, watch } from "vue";
 import { useI18n } from "vue-i18n";
 
 import HashLabel from "@/components/common/HashLabel.vue";
-import WalletInfoModal from "@/components/modals/WalletInfoModal.vue";
+import WalletInfoModal from "@/components/prividium/WalletInfoModal.vue";
 
 import useAddress from "@/composables/useAddress";
 import useContext from "@/composables/useContext";
