+++ packages/app/src/components/prividium/NetworkIndicator.vue
