+++ packages/app/src/components/prividium/WalletInfoModal.vue
